requirements-test.txt
-e .
flake8
isort
pytest
pytest-sugar
pytest-django
pytest-coverage
